# src/main/java/ar/edu/utn/frc/tup/lciii/dtos/common/CountryRequestDTO.java
package ar.edu.utn.frc.tup.lciii.dtos.common;


import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Setter
@Getter
@NoArgsConstructor
public class CountryRequestDTO {

    private Integer amountOfCountryToSave;
}